+++ example/cars_server/controller.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:math';
 
 import 'package:json_api/document.dart';
 import 'package:json_api/server.dart';
@@ -50,9 +51,16 @@ class CarsController implements JsonApiController {
     }
 
     if (res.toMany.containsKey(request.target.relationship)) {
+
+      final pageSize = 2;
+      final totalPages = max(0, res.toMany[request.target.relationship].length - 1) ~/ pageSize + 1;
+      final page = NumberedPage.fromQueryParameters(request.uri.queryParameters,
+          total: totalPages);
       final resources = res.toMany[request.target.relationship]
+          .skip(page.offset * pageSize)
+          .take(pageSize)
           .map((id) => _dao[id.type].fetchByIdAsResource(id.id));
-      return response.sendCollection(Collection(resources));
+      return response.sendCollection(Collection(resources, page: page));
     }
     return response
         .errorNotFound([JsonApiError(detail: 'Relationship not found')]);