+++ test/functional/fetch_test.dart
@@ -69,6 +69,17 @@ void main() async {
       expect(r.data.self.uri, uri);
     });
 
+    test('related collection travesal', () async {
+      final uri = Url.related('companies', '1', 'models');
+      final r0 = await client.fetchCollection(uri);
+      final firstPage = r0.data;
+      expect(firstPage.collection.length, 2);
+
+      final r1 = await client.fetchCollection(firstPage.pagination.last.uri);
+      final lastPage = r1.data;
+      expect(lastPage.collection.length, 2);
+    });
+
     test('404', () async {
       final r = await client.fetchCollection(Url.collection('unicorns'));
       expect(r.status, 404);